+++ temba/orgs/tests.py
@@ -2866,7 +2866,7 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
 
         self.login(self.user)
 
-        with self.assertNumQueries(13):
+        with self.assertNumQueries(12):
             response = self.client.get(home_url)
 
         # not so many options for viewers
@@ -2875,7 +2875,7 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
 
         self.login(self.admin)
 
-        with self.assertNumQueries(49):
+        with self.assertNumQueries(48):
             response = self.client.get(home_url)
 
         # more options for admins
@@ -2932,7 +2932,7 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
         menu_url = reverse("orgs.org_menu")
 
         self.assertMenu(menu_url, 8, ["Workspace/Child Workspace"])
-        self.assertMenu(f"{menu_url}settings/", 7)
+        self.assertMenu(f"{menu_url}settings/", 6)
 
         # agents should only see tickets and settings
         self.login(self.agent)
@@ -3011,7 +3011,7 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
             parent=self.org,
         )
 
-        with self.assertNumQueries(22):
+        with self.assertNumQueries(21):
             response = self.client.get(reverse("orgs.org_workspace"))
 
         # should have an extra menu option for our child (and section header)